增加判断条件

+++ update.ps1
@@ -1,13 +1,9 @@
-# 删除已有的文件
-Remove-Item -Path .\bucket -Recurse -Force
+if(Test-Path .\bucket) {
+    Remove-Item -Path .\bucket -Recurse -Force
+    New-Item -ItemType "directory" -Path ".\bucket"
+}
 
-# 将克隆的最新的文件拷贝到待处理的文件夹
-New-Item -ItemType "directory" -Path ".\bucket"
-
-# 仓库
 Copy-Item -Path ".\scoop-felixmaker\bucket\*" -Destination ".\bucket" -Recurse -Force
-
-# 复制完后，删掉克隆的文件夹
 Remove-Item -Path .\scoop-felixmaker -Recurse -Force
 
 Get-ChildItem -Recurse -Path .\bucket | ForEach-Object -Process {
